clean up error tables
