+++ mace/tools/tables_utils.py
@@ -20,7 +20,52 @@ def custom_key(key):
     return (2, key)
 
 
-def create_error_table(
+def create_error_tables(
+    train_valid_data_loader: dict,
+    test_data_loader: dict,
+    also_predict_committee: bool = False,
+    **kwargs
+):
+    table_train_valid = _create_error_table(
+            **kwargs,
+            all_data_loaders=train_valid_data_loader,
+            predict_committee=False,
+        )
+    if also_predict_committee:
+        logging.info("Error-table on TRAIN and VALID for single heads:\n" + str(table_train_valid))
+    else:
+        logging.info("Error-table on TRAIN and VALID:\n" + str(table_train_valid))
+
+    if also_predict_committee:
+        valid_loader = {"valid_committee": train_valid_data_loader["valid_committee-0"]}
+        table_train_valid = _create_error_table(
+            **kwargs,
+            all_data_loaders=valid_loader,
+            predict_committee=True,
+        )
+        logging.info("Error-table on VALID for whole committee:\n" + str(table_train_valid))
+
+    if test_data_loader:
+        table_test = _create_error_table(
+                **kwargs,
+                all_data_loaders=test_data_loader,
+                predict_committee=False
+            )
+        if also_predict_committee:
+            logging.info("Error-table on TEST for single heads:\n" + str(table_test))
+        else:
+            logging.info("Error-table on TEST:\n" + str(table_test))
+
+    if also_predict_committee and test_data_loader:
+        table_test = _create_error_table(
+            **kwargs,
+            all_data_loaders=test_data_loader,
+            predict_committee=True
+        )
+        logging.info("Error-table on TEST for whole committee:\n" + str(table_test))
+
+
+def _create_error_table(
     table_type: str,
     all_data_loaders: dict,
     model: torch.nn.Module,
@@ -29,6 +74,7 @@ def create_error_table(
     log_wandb: bool,
     device: str,
     distributed: bool = False,
+    predict_committee: bool = False,
 ) -> PrettyTable:
     if log_wandb:
         import wandb
@@ -108,6 +154,7 @@ def create_error_table(
             data_loader=data_loader,
             output_args=output_args,
             device=device,
+            predict_committee=predict_committee
         )
         if distributed:
             torch.distributed.barrier()
